Update index.html
Image created 6 April 2013

An image of thought and the considering of then-and-there bouts. And trying to work-out any sense of the observations at that time which I was making.
And now today, 13 Dcember 2025 there's so much arrived and going on that you can't shake-a-stick at it all.
And in frank parlance, this garbage needs to be sorted out. And soon.